requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
docutils==0.21.1
fonttools==4.51.0
idna==3.7
imageio==2.34.0
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.gstreamer==0.3.3
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
lazy_loader==0.4
matplotlib==3.8.4
networkx==3.3
numpy==1.26.4
packaging==24.0
pillow==10.3.0
Pygments==2.17.2
pyparsing==3.1.2
pypiwin32==223
python-dateutil==2.9.0.post0
pywin32==306
requests==2.31.0
scikit-image==0.23.1
scipy==1.13.0
six==1.16.0
tifffile==2024.2.12
urllib3==2.2.1
